A really big speed improvement. Essentially just fixes and simplifies code, allowing for a much faster server experience

+++ src/main/java/net/minecraft/src/EntityArrow.java
@@ -98,7 +98,7 @@ public class EntityArrow extends Entity implements IProjectile
 
     protected void entityInit()
     {
-        this.dataWatcher.addObject(16, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     /**
@@ -515,11 +515,11 @@ public class EntityArrow extends Entity implements IProjectile
 
         if (par1)
         {
-            this.dataWatcher.updateObject(16, Byte.valueOf((byte)(var2 | 1)));
+            this.dataWatcher.updateObject(16, (byte) (var2 | 1));
         }
         else
         {
-            this.dataWatcher.updateObject(16, Byte.valueOf((byte)(var2 & -2)));
+            this.dataWatcher.updateObject(16, (byte) (var2 & -2));
         }
     }
