+++ src/main/java/net/minecraft/src/EntityCreeper.java
@@ -67,8 +67,8 @@ public class EntityCreeper extends EntityMob
     protected void entityInit()
     {
         super.entityInit();
-        this.dataWatcher.addObject(16, Byte.valueOf((byte) - 1));
-        this.dataWatcher.addObject(17, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (byte) -1);
+        this.dataWatcher.addObject(17, (byte) 0);
     }
 
     /**
@@ -93,7 +93,7 @@ public class EntityCreeper extends EntityMob
     public void readEntityFromNBT(NBTTagCompound par1NBTTagCompound)
     {
         super.readEntityFromNBT(par1NBTTagCompound);
-        this.dataWatcher.updateObject(17, Byte.valueOf((byte)(par1NBTTagCompound.getBoolean("powered") ? 1 : 0)));
+        this.dataWatcher.updateObject(17, (byte) (par1NBTTagCompound.getBoolean("powered") ? 1 : 0));
 
         if (par1NBTTagCompound.hasKey("Fuse"))
         {
@@ -217,7 +217,7 @@ public class EntityCreeper extends EntityMob
      */
     public void setCreeperState(int par1)
     {
-        this.dataWatcher.updateObject(16, Byte.valueOf((byte)par1));
+        this.dataWatcher.updateObject(16, (byte) par1);
     }
 
     /**
@@ -226,6 +226,6 @@ public class EntityCreeper extends EntityMob
     public void onStruckByLightning(EntityLightningBolt par1EntityLightningBolt)
     {
         super.onStruckByLightning(par1EntityLightningBolt);
-        this.dataWatcher.updateObject(17, Byte.valueOf((byte)1));
+        this.dataWatcher.updateObject(17, (byte) 1);
     }
 }
